refactor: 增加 onRemovePanel 方法
