+++ src/Extensions/Components/BootstrapBlazor.DockView/wwwroot/js/dockview-utils.js
@@ -1,7 +1,7 @@
 ﻿import { DockviewComponent } from "./dockview-core.esm.js"
 import { DockviewPanelContent } from "./dockview-content.js"
 import { onAddGroup, addGroupWithPanel } from "./dockview-group.js"
-import { onAddPanel, getPanels, findPanelFunc, findPanel } from "./dockview-panel.js"
+import { onAddPanel, onRemovePanel, getPanels, findPanelFunc } from "./dockview-panel.js"
 import { getConfig, reloadFromConfig, loadPanelsFromLocalstorage } from './dockview-config.js'
 import './dockview-extensions.js'
 
@@ -45,45 +45,7 @@ const initDockview = (dockview, options, template) => {
     dockview.dispose = () => {
     }
 
-    dockview.onDidRemovePanel(event => {
-        let panel = {
-            id: event.id,
-            title: event.title,
-            component: event.view.contentComponent,
-            groupId: event.group.id,
-            params: {
-                ...event.params,
-                currentPosition: {
-                    width: event.group.element.parentElement.offsetWidth,
-                    height: event.group.element.parentElement.offsetHeight,
-                    top: parseFloat(event.group.element.parentElement.style.top || 0),
-                    left: parseFloat(event.group.element.parentElement.style.left || 0)
-                }
-            }
-        }
-        if (event.params.groupInvisible) {
-            panel.groupInvisible = event.params.groupInvisible
-        }
-        savePanel(dockview, panel)
-
-        // 在group上存储已删除的panel标识
-        !event.group.children && (event.group.children = [])
-        event.group.children = event.group.children.filter(p => findPanel(p, event) !== null);
-        event.group.children.push({
-            id: event.id,
-            title: event.title,
-            params: event.params
-        })
-
-        if (event.view.content.element) {
-            if (event.titleMenuEle) {
-                event.view.content.element.append(event.titleMenuEle)
-            }
-            if (dockview.params.template) {
-                dockview.params.template.append(event.view.content.element)
-            }
-        }
-    })
+    dockview.onDidRemovePanel(onRemovePanel);
 
     dockview.onDidAddPanel(onAddPanel);
